completed consistent styling of contact form (modal)
