--- src/components/pages/SupportPage/ContactUs.js
@@ -1,46 +1,0 @@
-import React from 'react';
-import emailjs from 'emailjs-com';
-import { Form, Input, Button } from 'antd';
-
-function ContactUs({ success, visible }) {
-  function sendEmail(e) {
-    e.preventDefault();
-    console.log(e.target);
-    emailjs
-      .sendForm(
-        process.env.REACT_APP_SERVICE_ID,
-        process.env.REACT_APP_TEMPLATE_ID,
-        e.target,
-        process.env.REACT_APP_USER_ID
-      )
-      .then(
-        result => {
-          console.log(result.text);
-        },
-        error => {
-          console.log(error.text);
-        }
-      );
-    e.target.reset();
-    success();
-    visible();
-  }
-
-  return (
-    <div>
-      <form className="Contact-Form-Container" onSubmit={sendEmail}>
-        <label>Name</label>
-        <input type="text" name="name" />
-        <label>Email</label>
-        <input type="email" name="email" />
-        <label>Subject</label>
-        <input type="text" name="subject" />
-        <label>Message</label>
-        <textarea name="message" />
-        <input type="submit" value="Send" />
-      </form>
-    </div>
-  );
-}
-
-export default ContactUs;
