+++ src/components/pages/SupportPage/SupportPageContainer.js
@@ -1,14 +1,12 @@
 import React, { useState } from 'react';
-import { Header } from '../../common';
-import ContactUs from './ContactUs';
-import { ToastContainer, toast } from 'react-toastify';
-import { Modal } from 'antd';
-import 'react-toastify/dist/ReactToastify.min.css';
+import RenderSupportPage from './RenderSupportPageModal';
+import { toast } from 'react-toastify';
 
 function SupportPageContainer(props) {
-  const [modalVisible, setModalVisible] = useState(false);
+  //this function takes care of the success message displayed on the screen
+  //when the user hits submit to send the form data to a dedicated email addy
+
   const toastifySuccess = () => {
-    console.log('works');
     toast('Form sent!', {
       position: 'bottom-right',
       autoClose: 3000,
@@ -20,41 +18,10 @@ function SupportPageContainer(props) {
       toastId: 'notifyToast',
     });
   };
+
   return (
     <>
-      <Header title="Support" displayMenu={true} />
-
-      <Modal
-        className="Contact-modal"
-        visible={modalVisible}
-        keyboard={true}
-        width={'100%'}
-        onCancel={() => setModalVisible(false)}
-        zIndex={2000}
-        cancelButtonProps={{ disabled: true }}
-        //contact us via our social here
-        footer="Information on Socials"
-        closeIcon="X"
-      >
-        <p>Contact Us</p>
-        <ContactUs
-          success={toastifySuccess}
-          visible={() => setModalVisible(false)}
-        />
-      </Modal>
-
-      <div className="support-page-container">
-        <button className="FAQ-button">
-          <h3>FAQ</h3>
-        </button>
-        <button
-          className="contact-button"
-          onClick={() => setModalVisible(true)}
-        >
-          <h3>Contact Us</h3>
-        </button>
-      </div>
-      <ToastContainer />
+      <RenderSupportPage success={toastifySuccess} />
     </>
   );
 }
