+++ src/styles/less/SupportPage.less
@@ -47,8 +47,7 @@
 
 //Contact Form Modal Styling
 
-.Contact-modal{
-
+.Contact-modal {
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -57,6 +56,47 @@
 .Contact-Form-Container {
   width: 50vw;
   display: flex;
-  flex-direction: column;
+  flex-flow: column nowrap;
   align-items: center;
+}
+
+.contact-form-inputs {
+  width: 60%;
+  margin-bottom: 15px;
+  letter-spacing: 5px;
+  font-weight: 600;
+  text-align: center;
+  border-radius: 4px;
+}
+
+.contact-form-message {
+  width: 60%;
+  height: 10vh;
+  font-weight: 300;
+  border-radius: 4px;
+}
+
+.contact-button-container {
+  display: flex;
+  flex-flow: column nowrap;
+  align-items: center;
+  width: 100%;
+  padding: 20px;
+}
+
+.contact-form-submit-button {
+  width: 50%;
+  border-radius: 14px;
+  justify-content: center;
+  cursor: pointer;
+  background: @button-blue;
+  font-size: 1rem;
+  background: @white-smoke;
+  border: 3px solid @button-blue;
+  color: @button-blue;
+  box-shadow: 2px 1px 1px;
+  &:hover {
+    background: @button-blue;
+    color: @white-smoke;
+  }
 }
